pyproject.toml
[tool.poetry]
name = "jwt-drf-passwordless"
version = "0.1.5"
description = "A passworldess plugin for Django Rest Framework authentication package"
authors = ["Sergio <smaisidoro@gmail.com>", "Mike Deeb <michael.f.deeb@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "jwt_drf_passwordless"}]

[tool.poetry.dependencies]
python = "^3.10"
django-sms = "^0.7.0"
djangorestframework-simplejwt = "^5.5.0"
djangorestframework = "^3.16.0"
django-templated-mail-2 = "^1.1.2"
django = "^5.2.4"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.1.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
tox = "^4.13.0"
pytest-django = "^4.9.0"
djet = "^0.3.0"
django = "^5.1.1"
coverage = "^7.6.1"
flake8 = "^7.1.1"
pytest-cov = "^5.0.0"
exceptiongroup = "^1.2.2"
ipykernel = "^6.29.5"
setuptools = "^75.1.0"
black = "^24.8.0"
twine = "^5.1.1"
sphinx = "^8.0.2"
watchdog = "^5.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"